requirements.txt
numpy==1.24.2
telegram==0.0.1
python-telegram-bot==20.2
tqdm==4.65.0
Pillow==9.5.0
moviepy==1.0.3
ffprobe-python==1.0.3
ffmpeg-python==0.2.0
replicate==0.7.0
requests==2.28.2
colorama==0.4.6
urllib3==1.26.7